BUGFIX
XE9.23
(RC23)

Fixed meta tags
Meta tags now have more extensive logging procedures to prevent future issues
Fixed an issue with efficiency values not being retrieved by saving and loading all unit names in learning files as lower-case
NTai no longer outputs to the game console if it cant find a units efficiency value
Updated RC22 strings to RC23
Changed the initial learning data generation to be more accurate, and so that it is generated if the learning data already exists
Modified numerous values to make learning algorithms more apparent.

due to inefficiencies in the AA build data for NTai partly caused by the many changes, i haven't updated the build trees and think that the AA build tree should be removed from the release branch if its in there(is it in there? Could someone check and remove it?)

git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@1974 37977431-3df6-0310-b722-df95706aa16b