*added crashDrag tag (float, default 0.005) to fighter/bomber planes, when the plane is crashing it uses this drag instead of the automatically calculated drag it uses when flying. The calculated drag is no longer used when crashing, even if no crashDrag tag is present in the unit fbi.
The default drag seems to be pretty much equal to the drag seen on TA-like planes that crash properly, the only deviations I've seen was when planes would crash super slowly because they have a maxAcc tag and I don't think anyone actually wanted that behaviour.

Realistically I think all the mods can just ignore the tag and use the default (which seems to be the most desired behaviour) but the tag is provided anyway in case a mod has any special plans.

git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@6690 37977431-3df6-0310-b722-df95706aa16b