UnitScript: revert bug introduced in e243496e5340b7e186d4aa1ab6f9502281a43414
This has to do with whether threads blocking on an animation are unblocked
or not before it's being replaced.  If an existing animation is replaced
by an animation of the same type, then, according to the comment on top
of the method, threads blocking on that animation will now wait for this
new turn/move instead.  If however a turn is replaced by a spin or
vice versa, then the original animation should be unblocked, otherwise
e.g. a wait-for-turn could wait infinitely until a spin stops spinning
(which it never does by itself)

The original code implemented this correctly; the bugged code implemented
the above for moves while always unblocking for turns.

As the comment above the method mentions, this behavior is arguable;
however it is still desired to use the same overriding behavior for both
moves and turns.

Hence this commit.