Proper algorithmic improvement to GetClosestEnemyUnit, GetClosestEnemyAircraft
The improvement is that instead of looping over CQuadField::Quad::units,
we loop over all allyteams and then over CQuadField::Quad::teamUnits.

This gives enormous improvement in the common case that not much enemies
are nearby, but many allies are. (As it doesn't loop at all over any of
the nearby allied units.)

(IOW, no need for a cache at all, this is more then twice as fast as cached
original algorithm, as it was in cbba8ae84efa9b37ffe1d67f350c75118f88f75c)

Conflicts:

	rts/Game/GameHelper.cpp