lua unit script: fixed Sleep/WaitFor not working in Killed
Couldn't yield through pcall. Added an 'onerror' callback function
on threads, which is called after a co_resume in which the thread
died due to an error. This is now used instead of pcall to ensure
SetDeathScriptFinished is called in Killed, even when an error occurs.

As a side effect stacktrace on error in Killed now works.