separate default build & HEADLESS_SYSTEM [CMake] * moved spring build to a separate builds subdir * introduce a new CMake config var HEADLESS_SYSTEM:BOOL * DS not built by default anymore
### sep. default build
This allows to separately define dependencies for the
different builds (default, GML, HL, DS).

### HEADLESS_SYSTEM
If TRUE:
- default and GML spring will not be configured
- spring-hl and spring-dedicated will be added to target "all"

This will allow configuring and building spring on a real headless machine,
without having to install the dependencies of normal spring.
So far, this is not yet possible, as we still depend on the OpenGL
and SDL headers with the headless stubs lubrary.