unitsync: FindFilesArchive: fix buffer overflow (CAUTION: this may break lobbies & autohosts!) Users of this function now have to indicate the nameBuf size. If the name-buffer supplied by the lobby is too small, an error is reported.
SpringLobby is already using this function as it works and is documented now.
TASClient & ZK-Lobby do not use the function (at least that is what i found
out by looking at their code-base).

If you use this function, and are unsure if you use it right, have a look
at the documentation in untisync_api.h