unuse AUX_SOURCE_DIRECTORY, following CMake doc recommendation [CMake] http://cmake.org/cmake/help/cmake-2-8-docs.html#command:aux_source_directory Most other noteworthy software i have seen that uses CMake is doing it this way too (without AUX_SOURCE_DIRECTORY).
Effects for devs:
* Each added/removed/moved native source file requires you to edit some
  CMakeLists.txt file.

Effects for everyone compiling spring:
* When a source-file gets added/removed/moved by someone else,
  you do not have to touch CMakeLists.txt anymore, to force re-scanning
  of source files.
This primarily applies when continuously updating sources through git.

This would (in theory) now also allow to do minor changes to the buildbot
script, so it would not force re-scanning of source files after every
commit. This in turn could save quite some time, especially with minor
commits, as they would not trigger recompiling of huge parts of the AI
subsystem anymore.