Remove "smooth" animations
That is, automatic conversion of a sequence of turn-now and move-now
followed by a sleep with 33 <= duration <= 300 in COB (not LUS!) to
turn and move with a speed calculated from the length of the sleep.

This should be fixed in the scripts and/or someone could write a
source-to-source or COB bytecode transformation that does pretty much
the same without the runtime performance impact.

The engine doesn't parse the smoothAnim UnitDef field anymore.