Remove data member introduced in eeb961d938bda7018d57050613f8e3d9a4759130 and rename chk->data
There is no reason to separately store the data and the checksum
when the checksum is the original simple XOR, which is equivalent
to an assignment checksum = data if size <= 4.

Added explicit if for the common case and renamed chk->data to make
this clearer. Might even be somewhat faster now, and in any case the
"checksums" displayed in the syncdebug log should now represent actual
int/float values again (at least most of the time).