add /ReloadShaders command (it reloads all engine shaders, not lua) what was needed to get this working: * had to move source code loading & defintions appending from CShaderHandler::CreateProgramObject to IShaderObject * had to write a GLSLCopyState(newProgID, oldProgID), it copies all GLSL shader relevant states (uniforms, ...) from one shaderprogram to another
future intentions:
* add runtime recompiling of uber-shaders: shaders have many definition flags to enable/disable features, a good example is the terrain shader with flags for shadows, SSMF, lights, ...
Such shaders have many flags and even more combinations of those (>50), compiling the shader in all combinations becomes more and more a problem. So I want to add a flag-state-machine.
Which allows you to toggle flags at runtime, and the state machine will recompile the shader if needed.