Revision to airdrag equation for SolidObject.cpp
In this revision:
Drag coefficients is multiplied with total speed (instead of component speed), but is split back into component value afterward.

This only make really tiny difference in unit trajectory, but is more accurate representation of air drag. Will not hurt to use it.

PS: I didn't test the code yet (except in LUA widget) because its hard to compile Spring in Windows.