Add support for SSMF Detail Normal Textures:
Works similarly to the previous SSMF splatting, but instead of sampling a combined texture of 4 different single-channel diffuse detail textures, it samples four separate normal textures.
The textures must contain the normal maps in the RGB channel (standard normal maps), with a diffuse luminance in the alpha  channel.
The system uses the same splatDistrTex as regular SSMF.

Usage:
	Specify splatDetailNormalTex1 .. splatDetailNormalTex4 in the MAP resources section.
	splatDetailTex will be unused if the above are present.