Add `GetUnitCurrentCommand` lua callout
Spring.GetUnitCurrentCommand(unitID) -> [cmdID, cmdOpts (coded), cmdTag[, cmdParam1[, cmdParam2 ... ]]]

Options are coded as a single number and parameters are returned on the stack so that there are no table allocations. This is a form of optimisation because getting the current command is a common action but so far the only available option was GetUnitCommands(unitID, 1) which still allocated 3 tables.