Prevent nullpo in LuaMenu Reload call (#481)
* Prevent nullpo in LuaMenu Reload call

`game` can be a null pointer when LuaMenu calls `SpringApp::Reload`.

In practice this didn't cause many problems, as this gets optimised as
a fixed function call, with a null `this` parameter. This led it to be
left unnoticed.

However, when using tools that are stricter about bad behaviour
(-fsanitize=...), this would detect a null pointer dereference and
abort the program.