fix loading a save made after reload under some circumstances.
The same C function can be registered from both LuaGaia and LuaRules. Then
in the save file it will be serialised under the (shorter) LuaGaia name.
The problem is that the C function registry is not cleaned every game,
therefore it might contain LuaGaia names while the map of the currently
saved has no LuaGaia. In that case if we load the game without having any
previous game with LuaGaia, the LuaGaia funcs won't be foundin the registry
and the game will crash.
This should be solved by clearing the registry at the end of every game.