Properly implement scancodes (#278)
* Revert "Fix reuse of actionOffset (#273)"

This reverts commit 941148f4935b82d1cb7cd007cd06f22d1c79f500.

* Revert "Replace KeyPressed in LuaMenuController by KeyPressedSC (#268)"

This reverts commit 7ff404a96fb65c8eda42f3871c7fbdd35459e42b.

* Revert "fix sdl (#257)"

This reverts commit e3d23aeebab099351d8f72d3f4de9e137e095568.

* Revert "Scancodes/Fixes for Compiling headless (#255)"

This reverts commit 87b53146bb8af4312f32c506eb999320c7604d2c.

* Revert "Capitalization fix for include "Game/UI/KeySetsc.h" -> SC"

This reverts commit 50c721653121d3a3113d607c8d39a563ca4e5fb7.

* Revert "Introducing SDL2 keyscancodes to Spring bindings :) (#236)"

This reverts commit 175f989b32238fd021e9870c43922ba01935172f.

* Add support for ScanCodes

- Deprecate DISALLOW_RELEASE_BINDINGS, it has not been in effect since
  2006. Up+ modifier is also deprecated, a warning is displayed if
  attempted to bind. No change in behavior
- KeyPress and KeyRelease callins receive an additional scanCode
  parameter
- sc_<k> keysets are introduced for scancodes and handled in conjunction
  with keycode bindings preserving the order on which the actions where
  bound. Any+ order is preserved (Any+ bindings are at the bottom)
- /keysave now preserves the binding order, providing a better output
- GetActionList(int keyCode=-1, int scanCode=-1, modifiers={alt=false,
  ctrl=false, shift=false, meta=false}) callout introduced to better
  provide query functionality for actions that would be triggered by a
  certain keyboard state
- widget, gadget and action Lua handlers are modified to support
  scancodes in `cont/`. Games that wish to support actions via scancodes
  are recommended to adopt these changes, otherwise behavior should be
  identical to before.

* Standardize key and scancode args signature

* Extract ordering preds

* Deduplicate actions with comments

As of 166806e91bbb7528fb4c182c67b5739ced280d1b we allowed actions
fallback when separate arguments provided but also didn't filter out
comments from the extra variable.

We remove comments and strip spaces from the right of extra now.

* Filter out duplicate code and scan bound actions

* Make order predicates explicit

* Use unordered_map instead of map

* Fix key and scancodes search not working correctly

* Use set instead of unordered_map

* Optimize actionlist merge for trigger order

* Fix misfit keychains returned to single keysets

* Cleaned up GetActionList

The main GetActionList now just returns a reference to the list of a
given CKeySet. Then it's up to the consumers to clean this reference up
- if they want to copy it, filter it, and so on.

Added two functions, filterByKeychain and filterByKeyset, which append
to a list the filtered actions from a given ActionList that match the
input keychain/keyset.
Ideally these at some point would become some wrappers around a const
ActionList &, offering some iterators, so we could just go over the
filtered actions without never having to copy them somewhere else.

Simplified the merge function so it assumes that the two input lists
have the same Any markers, and just appends the merged result to a list.

* Adjustments and fix stuff

* Extend GetKeyBindings instead of new callout

Extend GetKeyBindings to received an optional keyset to check for a
triggered code keyset and scan keyset instead of defining a new callout

Co-authored-by: Svalorzen <svalorzen@gmail.com>